MainProject/requirements.txt
Flask
selenium
requests
webdriver-manager
pandas
apache-airflow[amazon]